+++ src/main/groovy/org/codenarc/results/FileResults.groovy
@@ -73,12 +73,9 @@ class FileResults implements Results {
         new ArrayList(violations)
     }
 
-    /**
-     * Enable controlled access to the List of violations, allowing mutating that List
-     * @param closure the single-argument Closure; the argument is the List<Violation>
-     */
-    void withViolations(Closure closure) {
-        closure.call(violations)
+    /** Not intended for general use */
+    List getRawViolations() {
+        return violations
     }
 
     @Override